refactor(ApiKeyConfig): add useApiKeyConfig
