+++ frontend/app/config/components/index.tsx
@@ -1,3 +1,3 @@
-export * from "./ApiKeyConfig";
+export * from "./ApiKeyConfig/ApiKeyConfig";
 export * from "./ConfigForm";
 export * from "./ConfigTitle";
